requirements.txt
altair==4.2.0
asttokens 
attrs==22.2.0
backcall 
blinker==1.5
cachetools==5.3.0
certifi 
charset-normalizer==3.0.1
click==8.1.3
colorama 
comm 
contourpy==1.0.7
cycler==0.11.0
debugpy 
decorator 
entrypoints 
executing 
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
idna==3.4
importlib-metadata==6.0.0
ipykernel 
ipython 
jedi 
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client 
jupyter_core 
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline 
mdurl==0.1.2
mlxtend==0.21.0
nest-asyncio 
numpy==1.24.1
opencv-python==4.7.0.68
packaging 
pandas==1.5.3
parso 
pickleshare 
Pillow==9.4.0
platformdirs 
prompt-toolkit 
protobuf==3.20.3
psutil 
pure-eval 
pyarrow==10.0.1
pydeck==0.8.0
Pygments 
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil 
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
pyzmq 
requests==2.28.2
rich==13.2.0
scikit-learn==1.2.1
scipy==1.10.0
semver==2.13.0
six 
sklearn==0.0.post1
smmap==5.0.0
stack-data 
streamlit==1.17.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado 
traitlets 
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wcwidth 
wincertstore==0.2
zipp==3.11.0
